Moved utils to run_experiment
