+++ garage/experiment/experiment_wrapper.py
@@ -3,6 +3,9 @@ import argparse
 import ast
 import base64
 import datetime
+from enum import Enum
+import json
+import os
 import os.path as osp
 import pickle as pickle
 import signal
@@ -23,7 +26,6 @@ from garage.logger import logger
 from garage.logger import snapshotter
 from garage.logger.utils import dump_variant, log_parameters
 from garage.misc.console import colorize
-import garage.logger as logger
 import garage.plotter
 from garage.sampler import parallel_sampler
 from garage.sampler.utils import mask_signals
@@ -243,6 +245,42 @@ def child_proc_shutdown(children):
                       "of how the error was produced:\n")
         error_msg += ("https://github.com/rlworkgroup/garage/issues/120")
         print(colorize(error_msg, "yellow"))
+
+
+def log_parameters(log_file, args):
+    """Log parameters to file."""
+    log_params = {}
+    for param_name, param_value in args.__dict__.items():
+        log_params[param_name] = param_value
+    if args.args_data is not None:
+        log_params["json_args"] = dict()
+    mkdir_p(os.path.dirname(log_file))
+    with open(log_file, "w") as f:
+        json.dump(log_params, f, indent=2, sort_keys=True, cls=LogEncoder)
+
+
+def dump_variant(log_file, variant_data):
+    """Dump the variant file."""
+    mkdir_p(os.path.dirname(log_file))
+    with open(log_file, "w") as f:
+        json.dump(variant_data, f, indent=2, sort_keys=True, cls=LogEncoder)
+
+
+class LogEncoder(json.JSONEncoder):
+    """Encoder to be used as cls in json.dump."""
+
+    def default(self, o):
+        """Perform JSON encoding."""
+        if isinstance(o, type):
+            return {'$class': o.__module__ + "." + o.__name__}
+        elif isinstance(o, Enum):
+            return {
+                '$enum':
+                o.__module__ + "." + o.__class__.__name__ + '.' + o.name
+            }
+        elif callable(o):
+            return {'$function': o.__module__ + "." + o.__name__}
+        return json.JSONEncoder.default(self, o)
 
 
 if __name__ == "__main__":
