--- garage/logger/utils.py
@@ -1,45 +1,0 @@
-"""Contains utilities used in the logger.
-
-Misc methods, etc.
-"""
-from enum import Enum
-import json
-import os
-
-from garage.misc.console import mkdir_p
-
-
-def log_parameters(log_file, args):
-    """Log parameters to file."""
-    log_params = {}
-    for param_name, param_value in args.__dict__.items():
-        log_params[param_name] = param_value
-    if args.args_data is not None:
-        log_params["json_args"] = dict()
-    mkdir_p(os.path.dirname(log_file))
-    with open(log_file, "w") as f:
-        json.dump(log_params, f, indent=2, sort_keys=True, cls=LogEncoder)
-
-
-def dump_variant(log_file, variant_data):
-    """Dump the variant file."""
-    mkdir_p(os.path.dirname(log_file))
-    with open(log_file, "w") as f:
-        json.dump(variant_data, f, indent=2, sort_keys=True, cls=LogEncoder)
-
-
-class LogEncoder(json.JSONEncoder):
-    """Encoder to be used as cls in json.dump."""
-
-    def default(self, o):
-        """Perform JSON encoding."""
-        if isinstance(o, type):
-            return {'$class': o.__module__ + "." + o.__name__}
-        elif isinstance(o, Enum):
-            return {
-                '$enum':
-                o.__module__ + "." + o.__class__.__name__ + '.' + o.name
-            }
-        elif callable(o):
-            return {'$function': o.__module__ + "." + o.__name__}
-        return json.JSONEncoder.default(self, o)
